requirements.txt
boto3==1.35.81
PyJWT==2.9.0
scdl==2.11.3
mutagen==1.47.0
